requirements.txt
asgiref==3.5.0
coverage==6.3.2
Django==4.0.2
django-countries==7.3.2
django-debug-toolbar==3.2.4
django-js-asset==2.0.0
django-mptt==0.13.4
flake8==4.0.1
flake8-isort==4.1.1
isort==5.10.1
mccabe==0.6.1
modeltranslation==0.25
Pillow==9.0.1
pycodestyle==2.8.0
pyflakes==2.4.0
six==1.16.0
sqlparse==0.4.2
stripe3==0.0.0
testfixtures==6.18.5
typing-extensions==4.1.1
tzdata==2021.5
